refactor(rest): extract the ADR-0112 error/fault-classification prologue out of rest-server.ts
The module-level code ahead of `RestServer` is ADR-0112's concern — how a
thrown thing becomes an HTTP answer — not the REST server's. It reads no class
state and lived in that file for historical rather than structural reasons.

Moved verbatim to `packages/rest/src/error-response.ts`: mapDataError,
resolveErrorResponse, DATA_STORE_FAULT, UNCLASSIFIED_FAULT, isScriptFaultMessage,
declaredHttpStatus, truncateClientMessage, sendError, sendFieldVisibilityFault,
logWithheldServerFault, isExpectedDataStatus, isExpectedQueryRejection,
isExpectedRouteError, logUnexpectedRouteError, handleRouteError,
droppedFieldsHeaderValue, applyDroppedFieldsHeader, missingRelationIsObject,
plus their module constants. The console shim both files need moved to
`packages/rest/src/log.ts` rather than being copied.

No behaviour change: every moved line is byte-identical, the only edits being
six `export` prefixes and mapDataError's docblock reattached to the function it
describes (it had drifted above an unrelated import). `mapDataError` and
`zodIssuesToFields` are still exported from `./rest-server.js` unchanged.

Not the ADR-0076 D11 decomposition, which the 2026-08-15 ruling on #5949 closed.
